requirements.txt
# ============================================================================
# REQUIREMENTS COMPLETO - Dahell Intelligence
# ============================================================================

# --- Core Python Utilities ---
python-dotenv
requests

# --- Web Scraping ---
selenium
webdriver-manager

# --- Database ---
psycopg2-binary
SQLAlchemy

# --- Data Processing ---
numpy
pandas
pillow
psutil
openpyxl

# --- Machine Learning / AI ---
torch
torchvision
transformers
sentence-transformers
scikit-learn
sentencepiece>=0.1.99
protobuf>=3.20.0

# --- Web Framework (Django) ---
django>=5.0
django-environ
djangorestframework
django-cors-headers

# --- Utilities ---
tqdm
docker

# --- Market Intelligence ---
pytrends
pgvector
langchain
langchain-ollama
duckduckgo-search
tldextract
playwright
